ajustado calculo do valor de venda do pao

+++ desafio-aula-1/registradora-cheia-de-bugs/src/main/java/RelacaoPesoPreco.java
@@ -3,8 +3,8 @@ public class RelacaoPesoPreco {
     public static double retornaPrecoProduto(String item, int qtd) {
         double precoTotal = 0;
 
-        if ("pao".equals(item)) {
-            return precoTotal = 12.75 * (qtd * 60 / 1000);
+        if ("pao".equals(item)){
+            return precoTotal = 12.75 * ((double) qtd * 60 / 1000);
         }
 
         if ("torta".equals(item)) {
